Add unit test for failed lakectl file reads
Also, move the existing lakectl config test to the newly created config
test module.
