+++ tests/test_fs.py
@@ -1,6 +1,3 @@
-import sys
-
-from _pytest.logging import LogCaptureFixture
 from pytest import MonkeyPatch
 
 from lakefs_spec import LakeFSFileSystem
@@ -65,17 +62,3 @@ def test_initialization(monkeypatch: MonkeyPatch, temporary_lakectl_config: str)
     assert config.host == "http://lakefs.hello/api/v1"
     assert config.username == "my-user"
     assert config.password == "my-password"
-
-
-def test_lakectl_config_parsing_without_yaml(
-    monkeypatch: MonkeyPatch, caplog: LogCaptureFixture, temporary_lakectl_config: str
-) -> None:
-    LakeFSFileSystem.clear_instance_cache()
-
-    # unset YAML module from sys.modules.
-    monkeypatch.setitem(sys.modules, "yaml", None)
-    LakeFSFileSystem()
-
-    # look for the log record, assert that it has the `pyyaml` install instruction.
-    assert len(caplog.records) > 0
-    assert "`python -m pip install --upgrade pyyaml`." in caplog.records[0].message
